codeql fixes (#4)
* fix use of possibly unsafe "localtime"

* fix possible integer overflows

* add local codeql run script for Windows

* fix Linux/OSX build errors

* update codeql script docs

+++ Source/Core/AudioTypes/Source/OnsetDetector.cpp
@@ -488,11 +488,11 @@ void TOnsetDetector::CalculateOnsetFunction(const TOnsetFftProcessor::TPolarBuff
           
           if (Rectify)
           {
-            TotalDeviation += ::fabs(Deviation * TMathT<float>::Abs(pBuffer->mBin[i].mMagn));
+            TotalDeviation += ::fabs((double)Deviation * TMathT<float>::Abs(pBuffer->mBin[i].mMagn));
           }
           else
           {
-            TotalDeviation += ::fabs(Deviation);
+            TotalDeviation += ::fabs((double)Deviation);
           }
         }
       }
